refactor: enable bugprone-* in clang-tidy + fix narrowing conversions
After clearing the modernize-* / performance-* findings in #1432,
extended the .clang-tidy check set to include bugprone-* with two
exclusions for noise that doesn't fit the project's idioms:

- bugprone-throwing-static-initialization (67 findings): every
  static QString / QRegularExpression in the codebase. Standard Qt
  pattern; these constructors don't throw in practice. Disabling
  matches the convention in most Qt projects.

- bugprone-easily-swappable-parameters (22 findings): adjacent
  same-typed parameters. Stylistic; would require API-breaking
  signature changes that Qt's slot wiring expects.

That left 10 real bugprone-narrowing-conversions findings, all
qsizetype → int truncations (Qt 6 returns 64-bit qsizetype from
indexOf/size/count, but a lot of the codebase uses int).

Pattern: prefer qsizetype where the value stays internal; cast at
Qt-API boundaries that take int.

- configdialog.cpp setRowCount: static_cast<int> (Qt API)
- executor.cpp var.indexOf: int → qsizetype
- filecontent.cpp line.indexOf: int → qsizetype
- mainwindow.cpp tr count: static_cast<int> (Qt API)
- pass.cpp lastIndexOf: int → qsizetype
- pass.cpp *count = recipients.size(): static_cast<int> (out-param)
- passworddialog.cpp template indices: int → qsizetype
- qtpass.cpp prompt index: int → qsizetype

Validation:
- clang-tidy with new config: 0 findings in QtPass-owned code.
- tst_util: 106/106 pass.
- Full build clean.

cert-* yields no unique findings beyond bugprone aliases (the only
one was cert-err58-cpp = bugprone-throwing-static-initialization),
so not separately enabled.

+++ src/configdialog.cpp
@@ -575,7 +575,7 @@ void ConfigDialog::setProfiles(QHash<QString, QHash<QString, QString>> profiles,
   // Cache profiles for use in onProfileTableSelectionChanged
   m_profiles = profiles;
 
-  ui->profileTable->setRowCount(profiles.count());
+  ui->profileTable->setRowCount(static_cast<int>(profiles.count()));
   QHashIterator<QString, QHash<QString, QString>> i(profiles);
   int n = 0;
   while (i.hasNext()) {
